Week4/mtmc_dgnetpp_cityflow/requirements.txt
numpy>=1.24
scipy>=1.10
pandas>=2.0
tqdm>=4.66
pyyaml>=6.0
opencv-python>=4.8
torch>=2.1
torchvision>=0.16
scikit-learn>=1.3
matplotlib>=3.7
umap-learn>=0.5
Pillow>=4.2.1
lapsolver>=1.0.2
motmetrics>=1.1.3
pytrec_eval>=0.3
